Refactor imports to use the new Definition model and clean up related code
